build log & test error useContext()

+++ frontend/src/App.js
@@ -1,43 +1,49 @@
 // Formation OpenClassrooms - Développeur Web - Projet 7 - Grégory VENET
 
-import React, { useState, useEffect } from "react";
-import { BrowserRouter, Route, Routes } from "react-router-dom";
+import React, { useEffect, useState } from "react";
+import { BrowserRouter, Routes, Route, Navigate } from "react-router-dom";
 import Profil from "./pages/ProfilPage";
 import Home from "./pages/HomePage";
 import User from "./pages/UserPage";
 import Notify from "./pages/NotifPage";
 import Users from "./pages/UsersPage";
-// import { userIdContext } from "./components/AppContext";
+import { LoginContext } from "./components/AppContext";
 import { decodeToken, isExpired } from "react-jwt";
 
 const App = () => {
-	const [userId, setUserId] = useState(null);
-	const [token, setToken] = useState(null);
+	// let test = Navigate();
+	const [uId, setUid] = useState(null);
+	const [myToken, setMyToken] = useState(null);
 
-	useEffect(() => {
-		setUserId(parseInt(localStorage.getItem("id")));
-		setToken(localStorage.getItem("token"));
-		if (token && userId) {
-			const myDecodedToken = decodeToken(token);
-			const isMyTokenExpired = isExpired(token);
+	const isLogged = () => {
+		setUid(parseInt(localStorage.getItem("id")));
+		setMyToken(localStorage.getItem("token"));
+		if (myToken && uId) {
+			const myDecodedToken = decodeToken(myToken);
+			const isMyTokenExpired = isExpired(myToken);
 
-			if (!myDecodedToken || isMyTokenExpired) {
+			if (!myDecodedToken && isMyTokenExpired) {
 				localStorage.clear();
-				window.location = "/profil"; //a modifier
+				// test = "/profil";
 				return;
 			}
 		}
-	}, [token, userId]);
+	};
+
+	useEffect(isLogged);
 	return (
+		<LoginContext.Provider value={uId}>
 			<BrowserRouter>
 				<Routes>
-					<Route path="/profil" element={<Profil />} />
 					<Route path="/" element={<Home />} />
+					<Route path="/profil" element={<Profil />} />
 					<Route path="/users" element={<Users />} />
 					<Route path="/user" element={<User />} />
 					<Route path="/notify" element={<Notify />} />
+					<Route path="*" element={<Profil />} />
 				</Routes>
 			</BrowserRouter>
+		</LoginContext.Provider>
 	);
 };
